// src/app/modules/cpassapi/model/programma.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ente } from './ente';
import { Stato } from './stato';
import { Utente } from './utente';


export interface Programma { 
    id?: string;
    anno?: number;
    versione?: number;
    ente?: Ente;
    stato?: Stato;
    utenteReferente?: Utente;
    descrizione?: string;
    numeroProvvedimento?: number;
    descrizioneProvvedimento?: string;
    dataProvvedimento?: Date;
    dataPubblicazione?: Date;
    url?: string;
    codiceMit?: string;
    idRicevutoMit?: number;
    dataApprovazione?: Date;
    dataCreazione?: Date;
    utenteCreazione?: string;
    dataModifica?: Date;
    utenteModifica?: string;
    dataCancellazione?: Date;
    utenteCancellazione?: string;
    optlock?: string;
}